# docs/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: WeddingOS Unified API
  version: 1.0.0
  description: |
    Unified OpenAPI 3.0 specification for the WeddingOS microservices platform.

    This document was assembled from the actual service route files and related
    Zod / Prisma / Pydantic definitions in the monorepo. Paths are documented as
    gateway-facing routes using the service prefixes requested by the user
    (`/auth`, `/users`, `/vendors`, `/bookings`, `/payments`, `/execution`,
    `/notifications`, `/reviews`, `/search`, `/media`, `/chat`, `/ai`).

    Notes:
    - Most services return a common JSON envelope with `success`, `data`, and often `meta`.
    - Internal routes may accept either a bearer token with admin role or `x-internal-api-key`.
    - Chat real-time Socket.IO events are described in tag descriptions and README, while this spec focuses on HTTP endpoints.
servers:
  - url: http://localhost:8000
    description: Local API gateway
  - url: https://api-staging.weddingos.in
    description: Staging
  - url: https://api.weddingos.in
    description: Production
security:
  - bearerAuth: []
...
paths:
...
  /payments/order:
    post:
      tags: [Payments]
      summary: Create Razorpay order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
      responses:
        '201':
          description: Created payment order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentOrderEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
  /payments/verify:
    post:
      tags: [Payments]
      summary: Verify completed payment and capture funds
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyPaymentRequest'
      responses:
        '200':
          description: Verified payment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
  /payments/{id}/invoice:
    get:
      tags: [Payments]
      summary: Render payment invoice HTML
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: HTML invoice
          content:
            text/html:
              schema:
                type: string
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /payments/webhook:
    post:
      tags: [Payments]
      summary: Razorpay webhook
      description: Expects raw JSON body and `x-razorpay-signature` header.
      security: []
      parameters:
        - in: header
          name: x-razorpay-signature
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: Webhook accepted
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
        '400':
          description: Invalid signature
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Invalid signature
  /payments/{id}/refund:
    post:
      tags: [Payments]
      summary: Refund payment
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefundRequest'
      responses:
        '200':
          description: Refund result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefundEnvelope'
  /payments/escrow/{id}/release:
    post:
      tags: [Payments]
      summary: Release escrow to vendor
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Escrow released
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EscrowEnvelope'
  /payments/admin/list:
    get:
      tags: [Payments]
      summary: Admin list payments
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit100'
        - in: query
          name: status
          schema:
            $ref: '#/components/schemas/PaymentStatus'
      responses:
        '200':
          description: Paginated payment list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminPaymentListEnvelope'
...
  /payments/admin/disputes:
    get:
      tags: [Payments]
      summary: List disputes
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit100'
        - in: query
          name: status
          schema:
            $ref: '#/components/schemas/DisputeStatus'
      responses:
        '200':
          description: Paginated dispute list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DisputeListEnvelope'
    post:
      tags: [Payments]
      summary: Create dispute
      description: Route currently requires authentication but not explicit admin role middleware.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDisputeRequest'
      responses:
        '201':
          description: Created dispute
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DisputeEnvelope'
  /payments/admin/disputes/{id}/resolve:
    post:
      tags: [Payments]
      summary: Resolve dispute
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResolveDisputeRequest'
      responses:
        '200':
          description: Resolved dispute
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DisputeEnvelope'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    internalApiKey:
      type: apiKey
      in: header
      name: x-internal-api-key
  parameters:
    Page:
      in: query
      name: page
      schema:
        type: integer
        minimum: 1
        default: 1
    Limit100:
      in: query
      name: limit
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    IdPath:
      in: path
      name: id
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Validation or malformed request error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    Forbidden:
      description: Authenticated but not authorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
    ApiMeta:
      type: object
      properties:
        requestId:
          type: string
          nullable: true
        timestamp:
          type: string
          format: date-time
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          $ref: '#/components/schemas/ErrorObject'
        meta:
          $ref: '#/components/schemas/ApiMeta'
...
    PaymentStatus:
      type: string
      enum: [CREATED, PENDING, CAPTURED, FAILED, REFUNDED, PARTIALLY_REFUNDED]
...
    RefundReason:
      type: string
      enum: [CANCELLATION, DISPUTE_RESOLVED_CUSTOMER, OTHER]
    DisputeStatus:
      type: string
      enum: [OPEN, UNDER_REVIEW, RESOLVED_CUSTOMER, RESOLVED_VENDOR, CLOSED]
...
    CreateOrderRequest:
      type: object
      required: [bookingId, amountPaise, vendorId, eventDate]
      properties:
        bookingId: { type: string, format: uuid }
        amountPaise: { type: integer, minimum: 10000 }
        vendorId: { type: string, format: uuid }
        eventDate:
          type: string
          pattern: '^\d{4}-\d{2}-\d{2}$'
    VerifyPaymentRequest:
      type: object
      required: [razorpayOrderId, razorpayPaymentId, razorpaySignature, bookingId, eventDate]
      properties:
        razorpayOrderId: { type: string }
        razorpayPaymentId: { type: string }
        razorpaySignature: { type: string }
        bookingId: { type: string, format: uuid }
        eventDate:
          type: string
          pattern: '^\d{4}-\d{2}-\d{2}$'
    PaymentOrder:
      type: object
      properties:
        id: { type: string }
        amount: { type: integer }
        currency: { type: string }
        receipt: { type: string, nullable: true }
        status: { type: string, nullable: true }
        notes:
          type: object
          additionalProperties: true
          nullable: true
    Payment:
      type: object
      properties:
        id: { type: string, format: uuid }
        bookingId: { type: string, format: uuid }
        customerId: { type: string, format: uuid }
        vendorId: { type: string, format: uuid }
        amountPaise: { type: integer }
        currency: { type: string }
        description: { type: string, nullable: true }
        razorpayOrderId: { type: string }
        razorpayPaymentId: { type: string, nullable: true }
        status:
          $ref: '#/components/schemas/PaymentStatus'
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    EscrowHold:
      type: object
      properties:
        id: { type: string, format: uuid }
        paymentId: { type: string, format: uuid }
        bookingId: { type: string, format: uuid, nullable: true }
        vendorId: { type: string, format: uuid }
        status:
          $ref: '#/components/schemas/EscrowStatus'
        grossAmountPaise: { type: integer }
        platformFeePaise: { type: integer }
        gstOnFeePaise: { type: integer }
        vendorPayoutPaise: { type: integer }
        createdAt: { type: string, format: date-time }
        releasedAt: { type: string, format: date-time, nullable: true }
    RefundRequest:
      type: object
      properties:
        reason:
          $ref: '#/components/schemas/RefundReason'
        note:
          type: string
    PaymentOrderEnvelope:
      type: object
      properties:
        success: { type: boolean, example: true }
        data:
          type: object
          properties:
            order:
              $ref: '#/components/schemas/PaymentOrder'
        meta:
          $ref: '#/components/schemas/ApiMeta'
    PaymentEnvelope:
      type: object
      properties:
        success: { type: boolean, example: true }
        data:
          type: object
          properties:
            payment:
              $ref: '#/components/schemas/Payment'
        meta:
          $ref: '#/components/schemas/ApiMeta'
    EscrowEnvelope:
      type: object
      properties:
        success: { type: boolean, example: true }
        data:
          type: object
          properties:
            escrow:
              $ref: '#/components/schemas/EscrowHold'
        meta:
          $ref: '#/components/schemas/ApiMeta'
    RefundEnvelope:
      type: object
      properties:
        success: { type: boolean, example: true }
        data:
          type: object
          additionalProperties: true
        meta:
          $ref: '#/components/schemas/ApiMeta'
    AdminPaymentListEnvelope:
      type: object
      properties:
        success: { type: boolean, example: true }
        data:
          type: object
          properties:
            payments:
              type: array
              items:
                $ref: '#/components/schemas/Payment'
        meta:
          allOf:
            - $ref: '#/components/schemas/ApiMeta'
            - type: object
              properties:
                total: { type: integer }
                page: { type: integer }
                limit: { type: integer }
                pages: { type: integer }
...
    Dispute:
      type: object
      properties:
        id: { type: string, format: uuid }
        paymentId: { type: string, format: uuid }
        bookingId: { type: string, format: uuid, nullable: true }
        reason:
          $ref: '#/components/schemas/DisputeReason'
        status:
          $ref: '#/components/schemas/DisputeStatus'
        description: { type: string, nullable: true }
        resolutionNote: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    CreateDisputeRequest:
      type: object
      required: [paymentId, reason]
      properties:
        paymentId: { type: string, format: uuid }
        bookingId: { type: string, format: uuid }
        reason:
          $ref: '#/components/schemas/DisputeReason'
        description:
          type: string
    ResolveDisputeRequest:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [RESOLVED_CUSTOMER, RESOLVED_VENDOR, CLOSED]
        note:
          type: string
    DisputeEnvelope:
      type: object
      properties:
        success: { type: boolean, example: true }
        data:
          oneOf:
            - type: object
              properties:
                dispute:
                  $ref: '#/components/schemas/Dispute'
            - $ref: '#/components/schemas/Dispute'
        meta:
          $ref: '#/components/schemas/ApiMeta'
    DisputeListEnvelope:
      type: object
      properties:
        success: { type: boolean, example: true }
        data:
          type: object
          properties:
            disputes:
              type: array
              items:
                $ref: '#/components/schemas/Dispute'
        meta:
          allOf:
            - $ref: '#/components/schemas/ApiMeta'
            - type: object
              properties:
                total: { type: integer }
                page: { type: integer }
                limit: { type: integer }
                pages: { type: integer }
...